+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Axes: Added `AutoScaleExpand()` to zoom out to fit data only if necessary (#3259)
 * Style: Added `Plot.Style.ColorLegend()` for quick customization of legend colors (#3247)
 * Plot: Replacing palettes is now achieved by setting `Plot.Add.Palette` instead of `Plot.Palette`.
+* Plot: Added `ShowLegend()` overload that accepts manually created legend items
 
 ## ScottPlot 4.1.71 (not yet on NuGet)
 * Assembly: All ScottPlot packages use the same strong name signing key (#3235, #3241) _Thanks @mlessmann_